Feature: add support for a third Victron MPPT
only on ESP32-S3-USB. this fiddles with the available hardware UARTs to
make it possible to use a third Victron MPPT. if three MPPTs are defined
int the pin mapping, you will not be able to use the SmartShunt and JK
BMS battery interfaces.

note that using a second MPPT will also conflict with the SDM power
meter, and that conflict is not detected, yet.

+++ lib/VeDirectFrameHandler/VeDirectFrameHandler.cpp
@@ -62,7 +62,8 @@ VeDirectFrameHandler<T>::VeDirectFrameHandler() :
 }
 
 template<typename T>
-void VeDirectFrameHandler<T>::init(char const* who, int8_t rx, int8_t tx, Print* msgOut, bool verboseLogging, uint16_t hwSerialPort)
+void VeDirectFrameHandler<T>::init(char const* who, int8_t rx, int8_t tx,
+		Print* msgOut, bool verboseLogging, uint8_t hwSerialPort)
 {
 	_vedirectSerial = std::make_unique<HardwareSerial>(hwSerialPort);
 	_vedirectSerial->end(); // make sure the UART will be re-initialized
