+++ lib/VeDirectFrameHandler/VeDirectMpptController.h
@@ -40,7 +40,8 @@ class VeDirectMpptController : public VeDirectFrameHandler<veMpptStruct> {
 public:
     VeDirectMpptController() = default;
 
-    void init(int8_t rx, int8_t tx, Print* msgOut, bool verboseLogging, uint16_t hwSerialPort);
+    void init(int8_t rx, int8_t tx, Print* msgOut,
+        bool verboseLogging, uint8_t hwSerialPort);
 
     using data_t = veMpptStruct;
